API/requirement.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
astroid==3.3.8
async-timeout==5.0.1
attrs==24.3.0
autopep8==2.0.4
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dbus-python==1.2.18
deprecation==2.1.0
dill==0.3.9
distro==1.7.0
docstring-to-markdown==0.15
exceptiongroup==1.2.2
flake8==7.1.1
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.5.0
gotrue==2.11.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
MarkupSafe==3.0.2
mccabe==0.7.0
multidict==6.1.0
packaging==24.2
parso==0.8.4
platformdirs==4.3.6
pluggy==1.5.0
postgrest==0.19.1
propcache==0.2.1
pycodestyle==2.12.1
pydantic==2.10.5
pydantic_core==2.27.2
pydocstyle==6.3.0
pyflakes==3.2.0
Pygments==2.11.2
PyGObject==3.42.1
pylint==3.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.0
pytoolconfig==1.3.1
PyYAML==5.4.1
realtime==2.1.0
requests==2.32.3
rope==1.13.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
ssh-import-id==5.11
storage3==0.11.0
StrEnum==0.4.15
supabase==2.11.0
supafunc==0.9.0
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.3.0
websockets==13.1
Werkzeug==3.1.3
whatthepatch==1.0.7
yapf==0.43.0
yarl==1.18.3
